requirements.txt
opencv-python
lxml
tqdm
tensorflow-cpu
absl-py
easydict
matplotlib
pillow
flask

